fix some warning
x86 seems to have an diffrent header (ptrace.h)

+++ jni/ptraceaux.h
@@ -25,7 +25,7 @@ struct {
 	void (*ptrace_read_data) (pid_t pid, void *buf, void *addr, int nbytes);
 	int (*ptrace_strlen) (pid_t pid, void *addr);
 	void (*ptrace_write_data) (pid_t pid, void *buf, void *addr, int nbytes);
-	int (*ptrace_get_syscall_nr) (pid_t pid);
+	long (*ptrace_get_syscall_nr) (pid_t pid);
 	long (*ptrace_get_syscall_arg) (pid_t pid, int n);
 
 } ptrace_tool;
